requirements.txt
pandas>=2.0.0
numpy>=1.24.0
akshare>=1.15.0
yfinance>=0.2.0
vectorbt>=0.26.0
matplotlib>=3.7.0
mplfinance>=0.12.0
plotly>=5.0.0
pyyaml>=6.0
tqdm>=4.65.0
streamlit>=1.30.0
